Add "'cobalt_code': 1" to dom.gyp, fix Windows build
To make cobalt_code 1 work, the following changes are made:
- Added MSVC_PUSH_DISABLE_WARNING 4355 to anywhere "this" is used in base
  member initializer list.
- Disabled warning C4702: unreachable code in xtree.
- Added static_cast to all places where implicit conversion will cause warnings,
  or change variable type if it makes sense.
- Fixed all unreferenced parameter warnings.

b/22613037

+++ cobalt/dom/array_buffer.cc
@@ -26,9 +26,10 @@ ArrayBuffer::ArrayBuffer(uint32 length) : bytes_(length) {}
 scoped_refptr<ArrayBuffer> ArrayBuffer::Slice(int start, int end) {
   int clamped_start;
   int clamped_end;
-  ClampRange(start, end, byte_length(), &clamped_start, &clamped_end);
+  ClampRange(start, end, static_cast<int>(byte_length()), &clamped_start,
+             &clamped_end);
   scoped_refptr<ArrayBuffer> slice(
-      new ArrayBuffer(clamped_end - clamped_start));
+      new ArrayBuffer(static_cast<uint32>(clamped_end - clamped_start)));
   slice->bytes_.insert(slice->bytes_.begin(), bytes_.begin() + clamped_start,
                        bytes_.begin() + clamped_end);
   return slice;
